requirements.txt
anyascii==0.2.0
asgiref==3.3.4
beautifulsoup4==4.9.3
boto3==1.17.77
botocore==1.20.77
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dateparser==1.0.0
dj-database-url==0.5.0
Django==3.2.3
django-anymail==8.3
django-filter==2.4.0
django-modelcluster==5.1
django-storages==1.11.1
django-taggit==1.4.0
django-treebeard==4.5.1
django-webpack-loader==1.0.0
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
l18n==2020.6.1
MarkupSafe==2.0.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.1.1
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
wagtail==2.13
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
